Smart Invoice/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pandas==2.1.3
matplotlib==3.8.2
python-multipart==0.0.6
openpyxl==3.1.2